# [enhancement] improve performance of week() and yearweek()