--- a/be/src/vec/CMakeLists.txt
+++ b/be/src/vec/CMakeLists.txt
@@ -204,6 +204,7 @@ set(VEC_FILES
   runtime/vdatetime_value.cpp
   runtime/week_of_year.data
   runtime/week.data
+  runtime/year_week.data
   runtime/vdata_stream_recvr.cpp
   runtime/vdata_stream_mgr.cpp
   runtime/vfile_result_writer.cpp