# Proposed changes
ssb100,
doris_scanner_thread_pool_thread_num=1
sql: 
select  count(LO_ORDERDATE)  
FROM  lineorder  
WHERE  
     week(LO_ORDERDATE)  =  6;

week() improved from 19 sec to 15 sec 
yearweek() improved from 18 sec to 14 sec

Issue Number: close #xxx

## Problem Summary:

Describe the overview of changes.

## Checklist(Required)

1. Does it affect the original behavior: (Yes/No/I Don't know)
2. Has unit tests been added: (Yes/No/No Need)
3. Has document been added or modified: (Yes/No/No Need)
4. Does it need to update dependencies: (Yes/No)
5. Are there any changes that cannot be rolled back: (Yes/No)

## Further comments

If this is a relatively large or complex change, kick off the discussion at [dev@doris.apache.org](mailto:dev@doris.apache.org) by explaining why you chose the solution you did and what alternatives you considered, etc...