@englefly 2022-07-07
> 1. I think it's a really bad idea to initialize a look-up table from an explicit-defined data file. Could we do this job in BE process initialization phase? (e.g. doris_main.cpp)
A: using data file is better
-  it save launch time, meanwhile we have a bigger BE binary (80kb)
-  Since we do not need to add more source code to generate the table, saves maintain effort.

> 2. Using two look-up table with same keys also need to be re-thinking. Maybe a look-up table with mapping a key to a pair of value is enough?
A: the table dimension from high to low:mode->year->month->day  is better than that of year->month->day->mode, because former is more locality friendly.

> 3. I noticed there still are much unused codes. It seems using in your developing phase. Before merging to mater, it should be removed.
A: these codes are used to generate look-up table. As mention in comments, they are attached as comments to help us update the pre-calculated years. For example, if we want to pre-calculate year from 1970 to 2050.